Extract POSErrorState to a separate file
