+++ WooCommerce/Classes/POS/Models/PointOfSaleErrorState.swift
@@ -1,0 +1,31 @@
+import Foundation
+
+struct PointOfSaleErrorState: Equatable {
+    let title: String
+    let subtitle: String
+    let buttonText: String
+
+    static func errorOnLoadingProducts() -> Self {
+        PointOfSaleErrorState(title: Constants.failedToLoadTitle,
+                              subtitle: Constants.failedToLoadSubtitle,
+                              buttonText: Constants.failedToLoadButtonTitle)
+    }
+
+    enum Constants {
+        static let failedToLoadTitle = NSLocalizedString(
+            "pos.itemList.failedToLoadTitle",
+            value: "Error loading products",
+            comment: "Text appearing on the item list screen when there's an error loading products."
+        )
+        static let failedToLoadSubtitle = NSLocalizedString(
+            "pos.itemList.failedToLoadSubtitle",
+            value: "Give it another go?",
+            comment: "Text appearing on the item list screen as subtitle when there's an error loading products."
+        )
+        static let failedToLoadButtonTitle = NSLocalizedString(
+            "pos.itemList.failedToLoadButtonTitle",
+            value: "Retry",
+            comment: "Text for the button appearing on the item list screen when there's an error loading products."
+        )
+    }
+}
